requirements.txt
Flask
Flask-Cors
gunicorn
numpy
Pillow
tensorflow-cpu
pandas